content/canvas/test/webgl/non-conf-tests/test_webgl_request_context.html

<!DOCTYPE HTML>
<title>WebGL test: 'webgl' and 'experimental-webgl' context requests succeed,
    'moz-webgl' context requests fail.</title>
<script src="/MochiKit/MochiKit.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c1"></canvas>
<canvas id="c2"></canvas>
<script>

function testContextRetrieval(canvasId, creationId, shouldSucceed) {
  var canvas = document.getElementById(canvasId);
  ok(canvas, 'Invalid `canvasId`: ' + canvasId);

  var createdGL = canvas.getContext(creationId);
  if (shouldSucceed) {
    ok(createdGL, 'Request for \'' + creationId + '\' should succeed.');
  } else {
    ok(!createdGL, 'Request for \'' + creationId + '\' should fail.');
  }
}

testContextRetrieval('c1', 'experimental-webgl', true);
testContextRetrieval('c2', 'moz-webgl', false);

</script>
